+++ jest.config.js
@@ -6,6 +6,7 @@ module.exports = {
     roots: ['<rootDir>/src/'],
     moduleFileExtensions: ['js'],
     moduleDirectories: ['node_modules', 'src'],
+    setupFiles: ['<rootDir>/jest.setup.js'],
     testEnvironment: 'jsdom',
     verbose: false,
 };